requirements.txt
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
WTForms==3.0.1
